requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
databases==0.8.0
ecdsa==0.18.0
exceptiongroup==1.2.0
Farama-Notifications==0.0.4
fastapi==0.105.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
frozendict==2.3.10
fsspec==2023.12.1
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
gym==0.26.2
gym-anytrading==2.0.0
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
MetaTrader5==5.0.45
mpmath==1.3.0
multitasking==0.0.11
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==2.0.3
peewee==3.17.0
Pillow==10.1.0
protobuf==4.25.1
psycopg2==2.9.9
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pyspark==3.5.0
python-jose==3.3.0
pytz==2023.3.post1
pywin32==306
QuantStats==0.0.62
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
stable-baselines3==2.2.1
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
yfinance==0.2.33
